requirements.txt
# ==============================================================================
# DEPENDENCIAS DEL PROYECTO INVENTARIO UTP
# Actualizado: 2026-01-09
# ==============================================================================

# ------------------------------------------------------------------------------
# CORE DJANGO
# ------------------------------------------------------------------------------
Django>=4.2,<5.0

# ------------------------------------------------------------------------------
# BASE DE DATOS
# ------------------------------------------------------------------------------
psycopg2-binary==2.9.9    # PostgreSQL

# ------------------------------------------------------------------------------
# SERVIDOR WSGI (Producción)
# ------------------------------------------------------------------------------
gunicorn==21.2.0

# ------------------------------------------------------------------------------
# ARCHIVOS ESTÁTICOS
# ------------------------------------------------------------------------------
whitenoise==6.6.0         # Servir archivos estáticos sin Nginx

# ------------------------------------------------------------------------------
# CONFIGURACIÓN Y SEGURIDAD
# ------------------------------------------------------------------------------
python-decouple==3.8      # Variables de entorno

# ------------------------------------------------------------------------------
# UTILIDADES
# ------------------------------------------------------------------------------
Pillow==10.1.0            # Procesamiento de imágenes

# ------------------------------------------------------------------------------
# DESARROLLO (comentar en producción)
# ------------------------------------------------------------------------------
pytest==7.4.3
pytest-django==4.7.0
black==23.12.0
flake8==6.1.0
isort==5.13.2
coverage==7.3.2
